Use runes for client js
