+++ packages/integrations/svelte/package.json
@@ -23,7 +23,7 @@
     ".": "./dist/index.js",
     "./editor": "./dist/editor.cjs",
     "./*": "./*",
-    "./client.js": "./client.js",
+    "./client.js": "./client.svelte.js",
     "./server.js": {
       "default": "./server.js",
       "types": "./server.d.ts"
